+++ src/backends/octave/octavevariablemodel.cpp
@@ -33,6 +33,8 @@ void OctaveVariableModel::update()
         "      eval(['__cantor_string__ = disp(' __cantor_varname__ ');']);"
         "      printf([num2str(eval(['sizeof(' __cantor_varname__ ');'])) '\\n']);"
         "      printf([eval(['typeinfo(' __cantor_varname__ ');']) '\\n']);"
+        "      printf([num2str(eval(['rows(' __cantor_varname__ ');'])) '\\n']);"
+        "      printf([num2str(eval(['columns(' __cantor_varname__ ');'])) '\\n']);"
         "      printf(__cantor_string__);"
         "    catch"
         "      printf(['<unprintable value>' '\\n']);"
@@ -83,7 +85,7 @@ void OctaveVariableModel::parseNewVariables(Expression::Status status)
                 // every variable data has 4 parts/elements separated by a new line - the name of the variable, its size, type and the actual value
                 const auto& elements = data.split(QLatin1String("\n"), QString::SkipEmptyParts);
                 int count = elements.count();
-                if (count < 4)
+                if (count < 6)
                     continue;
 
                 const QString& name = elements.constFirst();
@@ -94,12 +96,14 @@ void OctaveVariableModel::parseNewVariables(Expression::Status status)
 
                 const QString& size = elements.at(1);
                 const QString& type = elements.at(2);
+                const QString& rows = elements.at(3);
+                const QString& columns = elements.at(4);
 
                 // the last section(s) contain(s) the value (single-line or multi-line output)
                 QString value;
                 if (OctaveSettings::self()->variableManagement())
                 {
-                    for (int i = 3; i < count; ++i)
+                    for (int i = 5; i < count; ++i)
                     {
                         if (!value.isEmpty())
                             value += QStringLiteral("; "); // separate multi-line values like in column vectors or in matrices by a semicolon
@@ -108,7 +112,7 @@ void OctaveVariableModel::parseNewVariables(Expression::Status status)
                 }
 
                 value = value.replace(QStringLiteral("   "), QStringLiteral(" ")); // for vectors Octave is separating the values with three blanks, replace with one
-                vars << Variable(name, value, size.toULongLong(), type);
+                vars << Variable(name, value, size.toULongLong(), type, rows + QStringLiteral("x") + columns);
             }
 
             setVariables(vars);